feat(YALB-1697): move spotlights JS to its own file
