+++ components/02-molecules/content-spotlight-portrait/yds-content-spotlight-portrait.twig
@@ -23,7 +23,6 @@
   'data-component-width': content_spotlight_portrait__width|default('site'),
   'data-component-theme': content_spotlight_portrait__theme|default('default'),
   'data-image-style': content_spotlight_portrait__style|default('inline'),
-  'data-remove-bottom-space': component__remove_bottom_space|default('false'),
   class: bem(content_spotlight_portrait__base_class)
 } %}
