+++ components/02-molecules/text-with-image/text-with-image.stories.js
@@ -7,8 +7,9 @@ import textWithImageTwig from './yds-text-with-image.twig';
 import imageData from '../../01-atoms/images/image/image.yml';
 import textWithImageData from './text-with-image.yml';
 
-const colorPairingsData = Object.keys(tokens['component-themes']);
+import '../content-spotlight-portrait/content-spotlights';
 
+const colorPairingsData = Object.keys(tokens['component-themes']);
 /**
  * Storybook Definition.
  */
