+++ components/02-molecules/text-with-image/yds-text-with-image.twig
@@ -23,7 +23,6 @@
   'data-image-position': text_with_image__position|default('image-left'),
   'data-component-width': text_with_image__width|default('site'),
   'data-component-theme': text_with_image__theme|default('default'),
-  'data-remove-bottom-space': component__remove_bottom_space|default('false'),
   class: bem(text_with_image__base_class)
 } %}
