+++ components/05-page-examples/standard-pages/standard-page-spotlights.twig
@@ -19,20 +19,17 @@
     {% include "@molecules/content-spotlight-portrait/yds-content-spotlight-portrait.twig" with {
       content_spotlight_portrait__style: 'inline',
       content_spotlight_portrait__theme: 'one',
-      component__remove_bottom_space: 'true',
     }%}
 
     {% include "@molecules/content-spotlight-portrait/yds-content-spotlight-portrait.twig" with {
         content_spotlight_portrait__style: 'inline',
         content_spotlight_portrait__position: 'image-right',
         content_spotlight_portrait__theme: 'two',
-        component__remove_bottom_space: 'true',
       }%}
 
     {% include "@molecules/content-spotlight-portrait/yds-content-spotlight-portrait.twig" with {
         content_spotlight_portrait__style: 'offset',
         content_spotlight_portrait__theme: 'one',
-        component__remove_bottom_space: 'true',
       }%}
 
     {% include "@molecules/content-spotlight-portrait/yds-content-spotlight-portrait.twig" with {
